+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2026-09-02
+
+### Added
+
+- Support for AppID
+- Support for Agent Intent
+
+### Changed
+
+- Remove unused "Privacy Report" screen
+
+## [1.2.2] - 2025-08-21
+
+### Changed
+
+- SDK update
+
+## [1.2.1] - 2025-10-29
+
+### Changed
+
+- Change 'Ledger Live' to 'Ledger Wallet'
+
 ## [1.2.0] - 2025-09-04
 
 ### Added